chore: controller/react types refactor
