+++ packages/controller/package.json
@@ -47,18 +47,18 @@
   "dependencies": {
     "@cartridge/controller-wasm": "catalog:",
     "@cartridge/penpal": "catalog:",
-    "micro-sol-signer": "^0.5.0",
-    "bs58": "^6.0.0",
-    "ethers": "^6.13.5",
     "@starknet-io/get-starknet-core": "5.0.0",
     "@starknet-io/types-js": "0.9.1",
-    "@wallet-standard/wallet": "^1.1.0",
     "@turnkey/sdk-browser": "^5.15.2",
+    "@wallet-standard/wallet": "^1.1.0",
+    "@walletconnect/ethereum-provider": "^2.20.0",
+    "bs58": "^6.0.0",
     "cbor-x": "^1.5.0",
-    "starknet": "^8.5.2",
+    "ethers": "^6.13.5",
+    "micro-sol-signer": "^0.5.0",
     "mipd": "^0.0.7",
     "open": "^10.1.0",
-    "@walletconnect/ethereum-provider": "^2.20.0"
+    "starknet": "^8.5.2"
   },
   "peerDependencies": {
     "react": "^18.2.0 || ^19.0.0",
@@ -75,24 +75,25 @@
   "devDependencies": {
     "@cartridge/controller-ui": "workspace:*",
     "@cartridge/tsconfig": "workspace:*",
-    "autoprefixer": "catalog:",
-    "postcss": "catalog:",
-    "tailwindcss": "catalog:",
+    "@microsoft/api-extractor": "^7",
     "@types/jest": "catalog:",
     "@types/mocha": "catalog:",
     "@types/node": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
+    "autoprefixer": "catalog:",
     "jest": "^29.7.0",
+    "postcss": "catalog:",
     "prettier": "catalog:",
     "react": "catalog:",
     "react-dom": "catalog:",
     "rollup-plugin-visualizer": "^5.12.0",
+    "tailwindcss": "catalog:",
     "ts-jest": "^29.2.5",
     "tsup": "catalog:",
     "typescript": "catalog:",
     "vite": "catalog:",
-    "vite-plugin-dts": "^4.5.3",
+    "vite-plugin-dts": "^5.0.3",
     "vite-plugin-node-polyfills": "^0.23.0",
     "vite-plugin-top-level-await": "catalog:",
     "vite-plugin-wasm": "catalog:"
