+++ packages/ui/src/components/primitives/toast/index.ts
@@ -1,6 +1,8 @@
-// Toast option types are the wire-protocol contract shared with keychain's
-// emitter (kept here). The toast UI components (Toast, useToast, the
-// specialized toasts and <ControllerToaster />) were moved to
-// `@/components/controller-react` and are published via the
-// `@cartridge/controller-ui/controller-react` subpath only.
-export * from "./types";
+// The toast wire-protocol contract lives with the receiver at
+// `@/components/controller-react/types` (single source of truth, also
+// published through `@cartridge/controller/react`). Re-exported here so the
+// keychain emitter keeps importing it from the `@cartridge/controller-ui`
+// root. The toast UI components (Toast, useToast, the specialized toasts and
+// <ControllerToaster />) live in `@/components/controller-react` and are
+// published via the `@cartridge/controller-ui/controller-react` subpath only.
+export * from "@/components/controller-react/types";
